chore: added logs for whitelabel script

+++ config_script/whitelabel.py
@@ -6,6 +6,7 @@ import subprocess
 import sys
 import yaml
 import json
+import coloredlogs
 
 class WhitelabelApp:
     def __init__(self, **kwargs):
@@ -48,14 +49,16 @@ class WhitelabelApp:
     def copy_assets(self):
         if self.assets:
             for asset in self.assets.items():
-                self.replace_images(asset[1])
-                self.replace_colors(asset[1])
+                self.replace_images(asset)
+                self.replace_colors(asset)
         else:
             logging.debug("Assets not found")
 
         
 
-    def replace_images(self, asset):
+    def replace_images(self, assetData):
+        asset = assetData[1]
+        assetName = assetData[0]
         if "images" in asset :
             assetPath = asset["imagesPath"] if "imagesPath" in asset else ""
             for name, image in asset["images"].items():
@@ -66,31 +69,61 @@ class WhitelabelApp:
                 darkImageName = image["darkImageName"] if "darkImageName" in image else ""
                 darkImageNameImport =  image["darkImageNameImport"] if "darkImageNameImport" in image else darkImageName
                 path_to_imageset = os.path.join(assetPath, currentPath, name+'.imageset')
-                # Delete current file(s)
+                # conditions to start updating
                 file_path = os.path.join(path_to_imageset, imageName)
-                if os.path.exists(file_path):
-                    os.remove(file_path)
+                dark_file_path = os.path.join(path_to_imageset, darkImageName)
+                files_to_changes_exist = os.path.exists(file_path) # 1
                 if hasDark:
-                    dark_file_path = os.path.join(path_to_imageset, darkImageName)
-                    if os.path.exists(dark_file_path):
-                        os.remove(dark_file_path)
-                # Change Contents.json
+                    files_to_changes_exist = files_to_changes_exist and os.path.exists(dark_file_path)
                 content_json_path = os.path.join(path_to_imageset, 'Contents.json')
-                with open(content_json_path, 'r') as openfile:
-                    contents_string = openfile.read()
-                contents_string = contents_string.replace(imageName, imageNameImport)
-                if hasDark:
-                    contents_string = contents_string.replace(darkImageName, darkImageNameImport)
-                with open(content_json_path, 'w') as openfile:
-                    openfile.write(contents_string)
-                # Copy new file(s)
+                contents_json_is_good = os.path.exists(content_json_path) # 2
+                if contents_json_is_good:
+                    with open(content_json_path, 'r') as openfile:
+                        contents_string = openfile.read()
+                    contents_json_is_good = contents_json_is_good and imageName in contents_string
+                    if hasDark:
+                        contents_json_is_good = contents_json_is_good and darkImageName in contents_string
+                
+                path_to_imageset_exists = os.path.exists(path_to_imageset) # 3
                 file_to_copy_path = os.path.join(self.assets_dir, imageNameImport)
-                shutil.copy(file_to_copy_path, path_to_imageset)
+                dark_file_to_copy_path = os.path.join(self.assets_dir, darkImageNameImport)
+                files_to_copy_exist = os.path.exists(file_to_copy_path) # 4
                 if hasDark:
-                    dark_file_to_copy_path = os.path.join(self.assets_dir, darkImageNameImport)
-                    shutil.copy(dark_file_to_copy_path, path_to_imageset)
+                    files_to_copy_exist = files_to_copy_exist and os.path.exists(dark_file_to_copy_path)
 
-    def replace_colors(self, asset):
+                if files_to_changes_exist and contents_json_is_good and path_to_imageset_exists and files_to_copy_exist:
+                    # Delete current file(s)
+                    os.remove(file_path)
+                    if hasDark:
+                        os.remove(dark_file_path)
+                    # Change Contents.json
+                    with open(content_json_path, 'r') as openfile:
+                        contents_string = openfile.read()
+                    contents_string = contents_string.replace(imageName, imageNameImport)
+                    if hasDark:
+                        contents_string = contents_string.replace(darkImageName, darkImageNameImport)
+                    with open(content_json_path, 'w') as openfile:
+                        openfile.write(contents_string)
+                    # Copy new file(s)
+                    shutil.copy(file_to_copy_path, path_to_imageset)
+                    logging.debug(assetName+"->images->"+name+": 'light mode'/universal image was updated with "+imageNameImport)
+                    if hasDark:
+                        shutil.copy(dark_file_to_copy_path, path_to_imageset)
+                        logging.debug(assetName+"->images->"+name+": 'dark mode' image was updated with "+darkImageNameImport)
+                else:
+                    # Handle errors
+                    if not files_to_changes_exist:
+                        logging.error(assetName+"->images->"+name+": original file(s) doesn't exist")
+                    elif not contents_json_is_good:
+                        logging.error(assetName+"->images->"+name+": Contents.json doesn't exist or wrong original file(s) in config")
+                    elif not path_to_imageset_exists:
+                        logging.error(assetName+"->images->"+name+": "+ path_to_imageset + " doesn't exist")
+                    elif not files_to_copy_exist:
+                        logging.error(assetName+"->images->"+name+": file(s) to copy doesn't exist")
+
+    def replace_colors(self, assetData):
+        asset = assetData[1]
+        assetName = assetData[0]
         if "colors" in asset:
             colorsPath = asset["colorsPath"] if "colorsPath" in asset else ""
             for name, color in asset["colors"].items():
@@ -100,20 +133,24 @@ class WhitelabelApp:
                 dark_color = color["dark"]
                 # Change Contents.json
                 content_json_path = os.path.join(path_to_colorset, 'Contents.json')
-                with open(content_json_path, 'r') as openfile:
-                    json_object = json.load(openfile)
-                    for key in range(len(json_object["colors"])):
-                        if "appearances" in json_object["colors"][key]:
-                            # dark
-                            changed_components = self.change_color_components(json_object["colors"][key]["color"]["components"], dark_color, name)
-                            json_object["colors"][key]["color"]["components"] = changed_components
-                        else:
-                            # light
-                            changed_components = self.change_color_components(json_object["colors"][key]["color"]["components"], light_color, name)
-                            json_object["colors"][key]["color"]["components"] = changed_components
-                    new_json = json.dumps(json_object) 
-                with open(content_json_path, 'w') as openfile:
-                    openfile.write(new_json)
+                if os.path.exists(content_json_path):
+                    with open(content_json_path, 'r') as openfile:
+                        json_object = json.load(openfile)
+                        for key in range(len(json_object["colors"])):
+                            if "appearances" in json_object["colors"][key]:
+                                # dark
+                                changed_components = self.change_color_components(json_object["colors"][key]["color"]["components"], dark_color, name)
+                                json_object["colors"][key]["color"]["components"] = changed_components
+                            else:
+                                # light
+                                changed_components = self.change_color_components(json_object["colors"][key]["color"]["components"], light_color, name)
+                                json_object["colors"][key]["color"]["components"] = changed_components
+                        new_json = json.dumps(json_object) 
+                    with open(content_json_path, 'w') as openfile:
+                        openfile.write(new_json)
+                    logging.debug(assetName+"->colors->"+name+": color was updated with light:'"+light_color+"' dark:'"+dark_color+"'")
+                else:
+                    logging.error(assetName+"->colors->"+name+": " + content_json_path + " doesn't exist")
 
     def change_color_components(self, components, color, name):
         color = color.replace("#", "")
@@ -136,7 +173,7 @@ def main():
     args = parser.parse_args()
 
     # DEBUG VARS
-    # args.config_file = "whitelabel.yaml"
+    # args.config_file = "../edx-mobile-config/openEdXAssets/whitelabel.yaml"
     # args.verbose = 2
 
     if args.help_config_file:
@@ -153,6 +190,9 @@ def main():
     if args.verbose > 1:
         log_level = logging.DEBUG
     logging.basicConfig(level=log_level)
+    logger = logging.getLogger(name='whitelabel_config')
+    coloredlogs.install(level=log_level, logger=logger)
+    # logger.propagate = False
 
     with open(args.config_file) as f:
         config = yaml.safe_load(f) or {}
